refactor: 주간/월간 집계 배치 파라미터 추가

+++ apps/commerce-batch/src/main/kotlin/com/loopers/domain/product/ProductMetricsRepository.kt
@@ -6,6 +6,6 @@ interface ProductMetricsRepository {
     fun findByMetricDateBetweenOrderByRankAscLimit(
         startDate: LocalDate,
         endDate: LocalDate,
-        limit: Int
+        limit: Long
     ): List<ProductMetrics>
 }
